refacto: storage code slicing
