+++ src/framework/modules/auth/actions.ts
@@ -10,7 +10,7 @@ import { Trackers } from '~/framework/util/tracker';
 import { clearRequestsCache } from '~/infra/fetchWithCache';
 import { OAuth2ErrorCode, destroyOAuth2, uniqueId, urlSigner } from '~/infra/oauth';
 
-import { StorageDict } from '~/framework/util/storage/storage';
+import { StorageSlice } from '~/framework/util/storage/slice';
 import {
   AuthError,
   ForgotMode,
@@ -236,7 +236,7 @@ function handleLoginRedirection(
     try {
       const { userData, userPublicInfo } = publicInfo;
       const sessionInfo = formatSession(platform, userInfo, userData, userPublicInfo, !!mustSaveSession);
-      await StorageDict.sessionInitAllStorages(sessionInfo);
+      await StorageSlice.sessionInitAllStorages(sessionInfo);
       if (partialSessionScenario) {
         const { defaultMobile, defaultEmail } = await getDefaultInfos(partialSessionScenario, platform.url);
         const context = await getAuthContext(platform);
